+++ src/main/java/com/mycompany/agency/Intern.java
@@ -3,18 +3,11 @@ package com.mycompany.agency;
 
 public class Intern extends StaffEmployee
 {
-    //-----------------------------------------------------------------
-    // Constructor: Sets up this intern using the specified information.
-    //-----------------------------------------------------------------
     public Intern(String eName, String eAddress, String ePhone)
     {
-        super(eName, eAddress, ePhone);
+        super(eName, eAddress, ePhone, "000-00-0000", 0.0);
     }
 
-    //-----------------------------------------------------------------
-    // Returns a zero pay value for this intern.
-    //-----------------------------------------------------------------
-    @Override
     public double pay()
     {
         return 0.0;
