+++ src/main/java/com/mycompany/agency/StaffMember.java
@@ -25,7 +25,8 @@ package com.mycompany.agency;
     //-----------------------------------------------------------------
     public String toString()
     {
-
+    return "Name: " + name + "\nAddress: " + address +
+            "\nPhone: " +  phone;
     }
 
     //-----------------------------------------------------------------
